refactor: rewrite with ToastManager
